\name{opentick-package}
\alias{opentick-package}
\alias{opentick}
\docType{package}
\title{
R API to opentick databases
}
\description{
Provide a native R interface to opentick real time and historical market data.\cr
Disclaimer: This software is not endorsed in any way by opentick corporation. The
authors of this software are not affiliated with opentick corporation.
}
\details{
\tabular{ll}{
Package: \tab opentick\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2008-09-27\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Josh Ulrich\cr
Maintainer: Josh Ulrich <josh.m.ulrich@gmail.com>
}
%\references{
%}
\keyword{ package }
%\seealso{
%}
%\examples{
%}
